refactor: Change structure & cleaning
