+++ src/test/api-loginCustomer.test.tsx
@@ -1,13 +1,13 @@
 import { ClientResponse, CustomerSignInResult } from '@commercetools/platform-sdk';
 import api from '@/api/api';
-import { apiRoot } from '@/api/commercetools-client';
+// import { apiRoot } from '@/api/commercetools-client';
 import { Mock } from 'vitest';
 
 describe('API Functions loginCustomer', () => {
   let mockExecute: Mock;
-  let mockPost: Mock;
-  let mockLogin: Mock;
-  let mockMe: Mock;
+  // let mockPost: Mock;
+  // let mockLogin: Mock;
+  // let mockMe: Mock;
 
   function createMockResponse(
     customer: CustomerSignInResult,
@@ -23,17 +23,16 @@ describe('API Functions loginCustomer', () => {
     vi.restoreAllMocks();
 
     mockExecute = vi.fn();
-    mockPost = vi.fn().mockReturnValue({
-      execute: mockExecute,
-    });
-    mockLogin = vi.fn().mockReturnValue({
-      post: mockPost,
-    });
-    mockMe = vi.fn().mockReturnValue({
-      login: mockLogin,
-    });
+    // mockPost = vi.fn().mockReturnValue({
+    //   execute: mockExecute,
+    // });
+    // mockLogin = vi.fn().mockReturnValue({
+    //   post: mockPost,
+    // });
+    // mockMe = vi.fn().mockReturnValue({
+    //   login: mockLogin,
+    // });
 
-    vi.spyOn(apiRoot.getApiRoot, 'me').mockImplementation(() => mockMe());
     vi.spyOn(api, 'logout').mockImplementation(() => Promise.resolve());
   });
 
@@ -64,9 +63,6 @@ describe('API Functions loginCustomer', () => {
     expect(result.signed).toBe(true);
     expect(result.message).toBe('OK');
     expect(result.response?.body.customer.email).toBe('test@example.com');
-    expect(mockPost).toHaveBeenCalledWith({
-      body: mockData,
-    });
     expect(api.logout).toHaveBeenCalled();
   });
 
@@ -76,42 +72,28 @@ describe('API Functions loginCustomer', () => {
       password: 'wrongpassword',
     };
 
-    mockExecute.mockResolvedValue(createMockResponse({} as CustomerSignInResult, 401));
-
-    const result = await api.loginCustomer(mockData);
-
-    expect(result.signed).toBe(false);
-    expect(result.message).toBe('Not Signed');
-    expect(result.response?.statusCode).toBe(401);
-  });
-
-  it('should handle API errors', async () => {
-    const mockData = {
-      email: 'test@example.com',
-      password: 'password123',
-    };
-
-    mockExecute.mockRejectedValue(new Error('Account with the given credentials not found.'));
+    mockExecute.mockResolvedValue(createMockResponse({} as CustomerSignInResult));
 
     const result = await api.loginCustomer(mockData);
 
     expect(result.signed).toBe(false);
     expect(result.message).toBe('Account with the given credentials not found.');
-    expect(result.response).toBeUndefined();
   });
 
-  it('should handle unknown errors', async () => {
+  it('should handle API errors', async () => {
     const mockData = {
       email: 'test@example.com',
-      password: 'password123',
+      password: 'password1234',
     };
 
-    mockExecute.mockRejectedValue({});
+    mockExecute.mockRejectedValue(
+      new Error('Customer account with the given credentials not found.')
+    );
 
     const result = await api.loginCustomer(mockData);
 
     expect(result.signed).toBe(false);
-    expect(result.message).toBe('Unknown error');
+    expect(result.message).toBe('Customer account with the given credentials not found.');
     expect(result.response).toBeUndefined();
   });
 });
