+++ tsconfig.json
@@ -2,7 +2,7 @@
   "files": [],
   "references": [
     { "path": "./tsconfig.app.json" },
-    { "path": "./tsconfig.node.json" },
+    // { "path": "./tsconfig.node.json" },
     { "path": "./tsconfig.vitest.json" }
   ]
 }
